// src/platform/packages/shared/kbn-connector-schemas/thehive/schemas/v1.ts
export const TheHiveConfigSchema = z
  .object({
    url: z.string(),
    organisation: z.string().nullable().default(null),
  })
  .strict();

export const TheHiveSecretsSchema = z
  .object({
    apiKey: z.string(),
  })
  .strict();

export const ExecutorSubActionPushParamsSchema = z
  .object({
    incident: z
      .object({
        title: z.string(),
        description: z.string(),
        externalId: z.string().nullable().default(null),
        severity: z.coerce.number().default(TheHiveSeverity.MEDIUM).nullable().default(null),
        tlp: z.coerce.number().default(TheHiveTLP.AMBER).nullable().default(null),
        tags: z.array(z.string()).nullable().default(null),
      })
      .strict(),
    comments: z
      .array(
        z
          .object({
            comment: z.string(),
            commentId: z.string(),
          })
          .strict()
      )
      .nullable()
      .default(null),
  })
  .strict();

export const PushToServiceIncidentSchema = {
  title: z.string(),
  description: z.string(),
  severity: z.coerce.number().nullable().default(null),
  tlp: z.coerce.number().nullable().default(null),
  tags: z.array(z.string()).nullable().default(null),
};

export const ExecutorSubActionGetIncidentParamsSchema = z
  .object({
    externalId: z.string(),
  })
  .strict();

export const ExecutorSubActionCreateAlertParamsSchema = z
  .object({
    title: z.string(),
    description: z.string(),
    type: z.string(),
    source: z.string(),
    sourceRef: z.string(),
    severity: z.coerce.number().default(TheHiveSeverity.MEDIUM).nullable().default(null),
    isRuleSeverity: z.boolean().default(false).nullable(),
    tlp: z.coerce.number().default(TheHiveTLP.AMBER).nullable().default(null),
    tags: z.array(z.string()).nullable().default(null),
    body: z.string().nullable().default(null),
  })
  .strict();

export const ExecutorParamsSchema = z.union([
  z
    .object({
      subAction: z.literal(SUB_ACTION.PUSH_TO_SERVICE),
      subActionParams: ExecutorSubActionPushParamsSchema,
    })
    .strict(),
  z
    .object({
      subAction: z.literal(SUB_ACTION.CREATE_ALERT),
      subActionParams: ExecutorSubActionCreateAlertParamsSchema,
    })
    .strict(),
]);